fix(core): eliminate residual heuristics and enforce precision
- Enforce float64 path augmentation in Kernel D
- Replace mode collapse threshold with config.entropy_threshold
- Use KernelType.N_KERNELS for weights and PRNG splits
- Add entropic penalty to Sinkhorn cost and keep converged as jnp.bool_
- Update core and kernels documentation PDFs

+++ stochastic_predictor/kernels/kernel_d.py
@@ -87,8 +87,8 @@ def create_path_augmentation(
         - Python.tex §2.2.4: Path Construction
     """
     n = signal.shape[0]
-    time_coords = jnp.arange(n, dtype=jnp.float32)
-    path = jnp.stack([time_coords, signal], axis=1)
+    time_coords = jnp.arange(n, dtype=jnp.float64)
+    path = jnp.stack([time_coords, signal.astype(jnp.float64)], axis=1)
     return path
